pyproject.toml
[tool.poetry]
name = "marimba"
version = "0.4.2"
description = "A Python framework for structuring, managing, processing and FAIR-ising scientific marine image datasets."
authors = [
    "Chris Jackett <chris.jackett@csiro.au>",
    "Kevin Barnard <kbarnard@mbari.org>"
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
typer = "^0.12.0"
rich = "^13.3.1"
PyYAML = "^6.0"
pandas = "^2.2.2"
ifdo = "^1.1.3"
pillow = "^10.3.0"
opencv-python = "^4.7.0.72"
gitpython = "^3.1.40"
boto3 = "^1.28.74"
staticmap = "^0.5.7"
piexif = "^1.1.3"
tabulate = "^0.9.0"
pyav = "^12.2.0"
distlib = "^0.3.8"
typing-extensions = "^4.12.2"

[tool.poetry.scripts]
marimba = "marimba.main:marimba_cli"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
isort = "^5.13.2"
pre-commit = "^3.5.0"
pytest = "^7.4.2"
flake8 = "^7.0.0"
mypy = "^1.10.0"
types-pyyaml = "^6.0.12.20240311"
bandit = "^1.7.8"
pydocstyle = "^6.3.0"
pylint = "^3.2.0"
pytest-cov = "^5.0.0"
pep8-naming = "^0.14.1"
flake8-bugbear = "^24.4.26"
flake8-comprehensions = "^3.14.0"
flake8-builtins = "^2.5.0"
types-tabulate = "^0.9.0.20240106"
types-requests = "^2.32.0.20240622"
ruff = "^0.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py310']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.env
  | \.venv
  | build
  | dist
  | temp
)/
'''
